refactor as utils
