+++ packages/xrpc/lib/src/utils.dart
@@ -1,0 +1,55 @@
+// Copyright 2023 Shinya Kato. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+import 'serializable.dart';
+
+dynamic removeNullValues(final dynamic object) {
+  if (object is Map) {
+    final parameters = <String, dynamic>{};
+    object.forEach((key, value) {
+      final newObject = removeNullValues(value);
+
+      if (newObject != null) {
+        parameters[key] = newObject;
+      }
+    });
+
+    return parameters.isNotEmpty ? parameters : null;
+  } else if (object is List) {
+    final parameters = <dynamic>[];
+    for (final value in object) {
+      final newObject = removeNullValues(value);
+
+      if (newObject != null) {
+        parameters.add(newObject);
+      }
+    }
+
+    return parameters.isNotEmpty ? parameters : null;
+  }
+
+  //! Just return it as is if it's neither Map nor List.
+  return object;
+}
+
+Map<String, dynamic> convertParameters(final Map<String, dynamic> parameters) =>
+    parameters.map((key, value) {
+      if (value is List?) {
+        return MapEntry(
+          key,
+          value?.map((e) => e.toString()).toList(),
+        );
+      } else if (value is Serializable) {
+        return MapEntry(
+          key,
+          value.value,
+        );
+      }
+
+      if (value is DateTime) {
+        return MapEntry(key, value.toUtc().toIso8601String());
+      }
+
+      return MapEntry(key, value.toString());
+    });
