+++ packages/xrpc/test/src/xrpc_test.dart
@@ -20,7 +20,6 @@ import 'package:xrpc/src/exception/invalid_request_exception.dart';
 import 'package:xrpc/src/exception/rate_limit_exceeded_exception.dart';
 import 'package:xrpc/src/exception/unauthorized_exception.dart';
 import 'package:xrpc/src/exception/xrpc_not_supported_exception.dart';
-import 'package:xrpc/src/serializable.dart';
 import 'package:xrpc/src/subscription.dart';
 import 'package:xrpc/src/xrpc.dart';
 import 'package:xrpc/src/xrpc_response.dart';
@@ -119,94 +118,6 @@ void main() {
         ),
         throwsA(isA<InternalServerErrorException>()),
       );
-    });
-  });
-
-  group('.removeNullValues', () {
-    test('simple map', () {
-      final actual = removeNullValues({
-        'test': 'aaaa',
-        'test2': null,
-      });
-
-      expect(actual, {'test': 'aaaa'});
-    });
-
-    test('nested map', () {
-      final actual = removeNullValues({
-        'test': 'aaaa',
-        'test2': {
-          'test': 'aaaa',
-          'test2': null,
-        },
-      });
-
-      expect(actual, {
-        'test': 'aaaa',
-        'test2': {
-          'test': 'aaaa',
-        }
-      });
-    });
-
-    test('list in map', () {
-      final actual = removeNullValues({
-        'test': 'aaaa',
-        'test2': [
-          {
-            'test': 'aaaa',
-            'test2': null,
-          },
-          null,
-        ],
-      });
-
-      expect(actual, {
-        'test': 'aaaa',
-        'test2': [
-          {
-            'test': 'aaaa',
-          }
-        ],
-      });
-    });
-  });
-
-  group('.convertParameters', () {
-    test('simple case', () {
-      final actual = convertParameters({'test': 'test'});
-
-      expect(actual, {'test': 'test'});
-    });
-
-    test('date time', () {
-      final now = DateTime.now();
-      final actual = convertParameters({'test': now});
-
-      expect(actual, {'test': now.toUtc().toIso8601String()});
-    });
-
-    test('date time', () {
-      final now = DateTime.now();
-      final actual = convertParameters({'test': now});
-
-      expect(actual, {'test': now.toUtc().toIso8601String()});
-    });
-
-    test('serializable enum', () {
-      final actual = convertParameters({'test': TestEnum.test});
-
-      expect(actual, {'test': 'test2'});
-    });
-
-    test('list', () {
-      final actual = convertParameters({
-        'test': [1, 2, 3]
-      });
-
-      expect(actual, {
-        'test': ['1', '2', '3']
-      });
     });
   });
 
@@ -547,13 +458,4 @@ void main() {
       }
     }, timeout: Timeout(Duration(minutes: 2)));
   });
-}
-
-enum TestEnum implements Serializable {
-  test('test2');
-
-  @override
-  final String value;
-
-  const TestEnum(this.value);
 }
